requirements.txt
Flask
python-dotenv
openai
chromadb
langchain
langchain-openai
langchain-community
langgraph
PyPDF2
python-docx
requests
Pillow
werkzeug
